requirements.txt
# 𝕽𝕺𝕏𝖄•𝔹𝕒𝕤𝕚𝕔ℕ𝕖𝕖𝕕𝔹𝕠𝕥 ⚡️
# Created by: RoxyBasicNeedBot
# Requirements file for Roxy Zip Maker Bot

pyrofork
TgCrypto
motor>=3.3.2
pymongo>=4.6.1
flask>=3.0.0
werkzeug>=3.0.1
requests>=2.31.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiohttp>=3.9.0
pyzipper>=0.3.6
psutil>=5.9.0
py7zr>=0.20.0

# 𝕽𝕺𝕏𝖄•𝔹𝕒𝕤𝕚𝕔ℕ𝕖𝕖𝕕𝔹𝕠𝕥 ⚡️
